+++ CHANGELOG.md
@@ -6,6 +6,7 @@ dev-master
 
 - [Elastic] Elasticsearch adapter
 - [Localization] Localization strategy support
+- [Config] Renamed `adapter_id` to `adapter` in configuration. See UPGRADE.md
 
 0.4.1
 -----